backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=11.3.0
opencv-python>=4.8.1.78
numpy>=1.26.0
torch>=2.1.0
torchvision>=0.16.0
scikit-image>=0.22.0